requirements.txt
streamlit
pandas
numpy
yfinance
matplotlib
requests
google-generativeai
pypdf
pandas_ta